+++ ydb/core/persqueue/pqtablet/common/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    event_helpers.cpp
     tracing_support.cpp
 )
 
@@ -8,7 +9,9 @@ SRCS(
 
 PEERDIR(
     ydb/core/base
+    ydb/core/persqueue/common
     ydb/library/logger
+    ydb/public/sdk/cpp/src/client/persqueue_public
 )
 
 END()
