+++ ydb/core/persqueue/ya.make
@@ -3,7 +3,6 @@ LIBRARY()
 SRCS(
     actor_persqueue_client_iface.h
     cluster_tracker.cpp
-    event_helpers.cpp
     heartbeat.cpp
     key.cpp
     list_all_topics_actor.cpp
